Cleanup headers
Remove array definitions from var.h and command.h, as they should really
not be there (makes it impossible to include multiple times). Instead
they now belong in their own .c objects.

+++ src/settings.c
@@ -19,7 +19,7 @@ void settings_read() {
 	settings_preset(settings.preset);
 }
 
-void settings_preset(int i) {
+void settings_preset(uint8_t i) {
 	eeprom_read_block(&preset, (void*)(sizeof(Settings)+i*sizeof(Preset)), sizeof(Preset));
 }
